Create Lazy Salesman
Ved is a salesman. He needs to earn at least W rupees in N days for his livelihood. However, on a given day i (1≤i≤N), he can only earn A[i] rupees by working on that day. Ved, being a lazy salesman, wants to take as many holidays as possible. It is known that on a holiday, Ved does not work and thus does not earn anything. Help maximize the number of days on which Ved can take a holiday. It is guaranteed that Ved can always earn at least W rupees by working on all days.